refactor(docs): avatar dx
