+++ apps/docs/content/components/avatar/bordered.raw.jsx
@@ -1,0 +1,14 @@
+import {Avatar} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4 items-center">
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a042581f4e29026024d" />
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258a2462d826712d" />
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a042581f4e29026704d" />
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026302d" />
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026702d" />
+      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026708c" />
+    </div>
+  );
+}
