+++ apps/docs/content/components/avatar/bordered.ts
@@ -1,17 +1,4 @@
-const App = `import {Avatar} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4 items-center">
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a042581f4e29026024d" />
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258a2462d826712d" />
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a042581f4e29026704d" />
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026302d" />
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026702d" />
-      <Avatar isBordered src="https://i.pravatar.cc/150?u=a04258114e29026708c" />
-    </div>
-  );
-}`;
+import App from "./bordered.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
