+++ apps/docs/content/components/avatar/custom.ts
@@ -1,18 +1,4 @@
-const App = `import {Avatar, AvatarIcon} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex items-center">
-      <Avatar
-        icon={<AvatarIcon />}
-        classNames={{
-          base: "bg-gradient-to-br from-[#FFB457] to-[#FF705B]",
-          icon: "text-black/80",
-        }}
-      />
-    </div>
-  );
-}`;
+import App from "./custom.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
