+++ apps/docs/content/components/avatar/fallbacks.ts
@@ -1,14 +1,4 @@
-const App = `import {Avatar} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4 items-center">
-      <Avatar showFallback src='https://images.unsplash.com/broken' />
-      <Avatar showFallback name='Jane' src='https://images.unsplash.com/broken' />
-      <Avatar name='Joe' src='https://images.unsplash.com/broken' />
-    </div>
-  );
-}`;
+import App from "./fallbacks.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
