+++ apps/docs/content/components/avatar/group-custom-count.ts
@@ -1,24 +1,4 @@
-const App = `import {Avatar, AvatarGroup} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <AvatarGroup
-      isBordered
-      max={3}
-      total={10}
-      renderCount={(count) => (
-        <p className="text-small text-foreground font-medium ms-2">+{count} others</p>
-      )}
-    >
-      <Avatar src="https://i.pravatar.cc/150?u=a042581f4e29026024d" />
-      <Avatar src="https://i.pravatar.cc/150?u=a04258a2462d826712d" />
-      <Avatar src="https://i.pravatar.cc/150?u=a042581f4e29026704d" />
-      <Avatar src="https://i.pravatar.cc/150?u=a04258114e29026302d" />
-      <Avatar src="https://i.pravatar.cc/150?u=a04258114e29026702d" />
-      <Avatar src="https://i.pravatar.cc/150?u=a04258114e29026708c" />
-    </AvatarGroup>
-  );
-}`;
+import App from "./group-custom-count.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
