+++ apps/docs/content/components/avatar/group-custom-impl.raw.jsx
@@ -1,0 +1,29 @@
+import {forwardRef} from "react";
+import {Avatar, useAvatarGroup, AvatarGroupProvider} from "@nextui-org/react";
+
+const AvatarGroup = forwardRef((props, ref) => {
+  const {
+    Component,
+    clones,
+    context,
+    remainingCount,
+    renderCount = (count) => <Avatar name={`+${count}`} />,
+    getAvatarGroupProps,
+  } = useAvatarGroup({
+    ref,
+    ...props,
+  });
+
+  return (
+    <Component {...getAvatarGroupProps()}>
+      <AvatarGroupProvider value={context}>
+        {clones}
+        {remainingCount > 0 && renderCount(remainingCount)}
+      </AvatarGroupProvider>
+    </Component>
+  );
+});
+
+AvatarGroup.displayName = "AvatarGroup";
+
+export default AvatarGroup;
