+++ apps/docs/content/components/avatar/group-custom-impl.raw.tsx
@@ -1,0 +1,36 @@
+import {forwardRef} from "react";
+import {
+  Avatar,
+  useAvatarGroup,
+  AvatarGroupProvider,
+  AvatarGroupProps as BaseAvatarGroupProps,
+} from "@nextui-org/react";
+
+export interface AvatarGroupProps extends BaseAvatarGroupProps {}
+
+const AvatarGroup = forwardRef<HTMLDivElement, AvatarGroupProps>((props, ref) => {
+  const {
+    Component,
+    clones,
+    context,
+    remainingCount,
+    renderCount = (count) => <Avatar name={`+${count}`} />,
+    getAvatarGroupProps,
+  } = useAvatarGroup({
+    ref,
+    ...props,
+  });
+
+  return (
+    <Component {...getAvatarGroupProps()}>
+      <AvatarGroupProvider value={context}>
+        {clones}
+        {remainingCount > 0 && renderCount(remainingCount)}
+      </AvatarGroupProvider>
+    </Component>
+  );
+});
+
+AvatarGroup.displayName = "AvatarGroup";
+
+export default AvatarGroup;
