+++ apps/docs/content/components/avatar/group-custom-impl.ts
@@ -1,68 +1,5 @@
-const App = `import {forwardRef} from "react";
-
-import {useAvatarGroup, AvatarGroupProvider} from "@nextui-org/react";
-import Avatar from "./avatar"; // your custom avatar or nextui avatar
-
-const AvatarGroup = forwardRef((props, ref) => {
-  const {
-    Component,
-    clones,
-    context,
-    remainingCount,
-    renderCount = (count) => <Avatar name={\`+\${count}\`} />,
-    getAvatarGroupProps,
-  } = useAvatarGroup({
-    ref,
-    ...props,
-  });
-
-  return (
-    <Component {...getAvatarGroupProps()}>
-      <AvatarGroupProvider value={context}>
-        {clones}
-        {remainingCount > 0 && renderCount(remainingCount)}
-      </AvatarGroupProvider>
-    </Component>
-  );
-});
-
-AvatarGroup.displayName = "AvatarGroup";
-
-export default AvatarGroup;`;
-
-const AppTs = `import {forwardRef} from "react";
-
-import {useAvatarGroup, AvatarGroupProvider, AvatarGroupProps as BaseAvatarGroupProps} from "@nextui-org/react";
-import Avatar from "./avatar"; // your custom avatar or nextui avatar
-
-export interface AvatarGroupProps extends BaseAvatarGroupProps {}
-
-const AvatarGroup = forwardRef<HTMLDivElement, AvatarGroupProps>((props, ref) => {
-  const {
-    Component,
-    clones,
-    context,
-    remainingCount,
-    renderCount = (count) => <Avatar name={\`+\${count}\`} />,
-    getAvatarGroupProps,
-  } = useAvatarGroup({
-    ref,
-    ...props,
-  });
-
-  return (
-    <Component {...getAvatarGroupProps()}>
-      <AvatarGroupProvider value={context}>
-        {clones}
-        {remainingCount > 0 && renderCount(remainingCount)}
-      </AvatarGroupProvider>
-    </Component>
-  );
-});
-
-AvatarGroup.displayName = "AvatarGroup";
-
-export default AvatarGroup;`;
+import App from "./group-custom-impl.raw.jsx?raw";
+import AppTs from "./group-custom-impl.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
